+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 8.1
 - Fix Conflict Warning displayed twice in Courses.php
 - Fix PHP Notice Undefined index in miscExport.fnc.php
 - Fix SQL error when Student / Staff ID is hacked / not an integer in URL in Current.php
+- SQL ACCOUNTING_SALARIES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
 
 Changes in 8.0.4
 ----------------